Restored mixin based implementation of `observer` for class based components
